add test profile

+++ backend/src/main/java/hw/spring/services/activity/ActivityDefaultService.java
@@ -10,23 +10,28 @@ import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.Optional;
 
-/**
- * Created by Kamil on 31-Mar-17.
- */
+
 @Service
 public class ActivityDefaultService implements ActivityService {
 
     private static final int MAX_ACTIVITIES = 8;
-    @Inject ActivityRepository activityRepository;
+    @Inject
+    ActivityRepository activityRepository;
+
+    private static boolean hasntActivityId(Activity activity) {
+        return activity.getId() == 0;
+    }
 
     public void setActivityRepository(ActivityRepository activityRepository) {
         this.activityRepository = activityRepository;
     }
 
+    @Override
     public List<Activity> getAll() {
         return activityRepository.findAllByOrderById();
     }
 
+    @Override
     public Optional<Activity> getById(int id) {
         return activityRepository.findById(id);
     }
@@ -35,31 +40,23 @@ public class ActivityDefaultService implements ActivityService {
     public void addActivity(Activity activity) throws BadRequestException {
         assert null != activityRepository;
 
-        if (activityRepository.count() > MAX_ACTIVITIES) {
-            throw new BadRequestException("Max activities count reached");
-        }
+        if (activityRepository.count() > MAX_ACTIVITIES) throw new BadRequestException("Max activities count reached");
 
         activityRepository.save(activity);
     }
 
-    private boolean hasntActivityId(Activity activity) {
-        return activity.getId() == 0;
+    @Override
+    public void deleteActivity(int id) {
+        if (activityRepository.existsById(id)) activityRepository.deleteById(id);
     }
 
-    public void deleteActivity(int id) {
-        if (activityRepository.existsById(id)) {
-            activityRepository.deleteById(id);
-        }
+    @Override
+    public void updateActivity(int id, Activity activity) {
+        if (null == activity) throw new NoSuchElementException();
+        activityRepository.save(activity);
     }
 
     public void deleteAll() {
         activityRepository.deleteAll();
-    }
-
-    public void updateActivity(int id, Activity activity) {
-        if(null == activity) {
-            throw new NoSuchElementException();
-        }
-        activityRepository.save(activity);
     }
 }
